Docker image for building gradle projects with gradle 2.9 and Java 8
The image defines the volume /project that must be mapped to the host
directory containing the gradle project.

The user can also specify custom uid and gid to perform the build,
otherwise the build artifacts generated in the host filesystem will
be owned by root.

The JVM's user.home property is bound to the GRADLE_HOME environment
variable. The GRADLE_HOME folder (/gradle) allows read and write access
to any user.